requirements.txt
discord.py==1.7.3
pymongo==3.12.0
asyncio==3.4.3
aiohttp==3.7.4
dnspython==2.1.0
python-dotenv==0.18.0
aiohttp==3.7.4.post0
html==1.16